backend-server/requirements-worker.txt
# Core dependencies for simulation workers
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.7.1
typing_extensions==4.12.2

# Database
SQLAlchemy==2.0.38
sqlmodel==0.0.22
psycopg2-binary==2.9.10

# Docker management
docker==7.1.0

# RabbitMQ
aio-pika

# Storage
supabase

# Cache
redis

# Utilities
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# Async support
anyio==4.8.0
sniffio==1.3.1
exceptiongroup==1.2.2 